+++ functions/public/Complete-PSWorkItem.ps1
@@ -33,16 +33,21 @@ Function Complete-PSWorkItem {
         [Switch]$PassThru
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand): Starting"
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand): PSBoundParameters"
-        $PSBoundParameters | Out-String | Write-Verbose
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand): Opening a connection to $Path"
+        $PSDefaultParameterValues["_verbose:Command"] = $MyInvocation.MyCommand
+        $PSDefaultParameterValues["_verbose:block"] = "Begin"
+        _verbose -message $strings.Starting
+        _verbose -message ($strings.UsingDB -f $path)
+        #Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand): Starting"
+        Write-Debug "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand): PSBoundParameters"
+        $PSBoundParameters | Out-String | Write-Debug
+        #Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand): Opening a connection to $Path"
         Try {
             $conn = Open-MySQLiteDB -Path $Path -ErrorAction Stop
             $conn | Out-String | Write-Debug
         }
         Catch {
-            Throw "$($MyInvocation.MyCommand): Failed to open the database $Path"
+            #Throw "$($MyInvocation.MyCommand): Failed to open the database $Path"
+            Throw "$($MyInvocation.MyCommand): $($strings.FailToOpen -f $Path)"
         }
 
         #parameters to splat to Invoke-MySQLiteQuery
@@ -55,11 +60,14 @@ Function Complete-PSWorkItem {
     } #begin
 
     Process {
-        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Completing task id $ID "
+        $PSDefaultParameterValues["_verbose:block"] = "Process"
+        #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Completing task id $ID "
+        _verbose -message ($strings.CompleteTask -f $ID)
         #test if archive table has been updated to include the OriginalTaskID column
         $test = Invoke-MySQLiteQuery -Path $path -query "pragma table_info('archive')" | Where-Object name -eq 'id'
         if (-Not $test) {
-            Write-Warning "Cannot verify the archive table column ID. Please run Update-PSWorkItemDatabase to update the table then try completing the command again. It is recommended that you backup your database before updating the table."
+            Write-Warning $strings.FailedArchiveID
+            #"Cannot verify the archive table column ID. Please run Update-PSWorkItemDatabase to update the table then try completing the command again. It is recommended that you backup your database before updating the table."
             Return
         }
         #validate the task id
@@ -69,7 +77,8 @@ Function Complete-PSWorkItem {
             $task = Invoke-MySQLiteQuery @splat
         }
         Catch {
-            Write-Warning "Failed to execute query $($splat.query)"
+            Write-Warning ($strings.FailedQuery -f $splat.query)
+            #"Failed to execute query $($splat.query)"
             Close-MySQLiteDB $conn
             Throw $_
         }
@@ -81,52 +90,60 @@ Function Complete-PSWorkItem {
                     Invoke-MySQLiteQuery @splat
                 }
                 Catch {
-                    Write-Warning "Failed to execute query $($splat.query)"
+                    Write-Warning ($strings.FailedQuery -f $splat.query)
+                    #"Failed to execute query $($splat.query)"
                     Close-MySQLiteDB $conn
                     Throw $_
                 }
                 #copy the task to the archive table
                 $splat.query = "INSERT into archive SELECT * from tasks WHERE ID= '$ID'"
                 # "INSERT into archive SELECT *,ROWID AS originalid from tasks WHERE RowID= '$ID'"
-                Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Moving item to Archive."
+                #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Moving item to Archive."
+                _verbose -message $strings. MoveItem
                 Write-Debug $splat.query
                 Try {
                     Invoke-MySQLiteQuery @splat
                 }
                 Catch {
-                    Write-Warning "Failed to execute query $($splat.query)"
+                    Write-Warning ($strings.FailedQuery -f $splat.query)
+                    #"Failed to execute query $($splat.query)"
                     Close-MySQLiteDB $conn
                     Throw $_
                 }
 
                 #Validate the copy using the task GUID
                 $splat.query = "SELECT * from archive WHERE taskid='$($task.taskid)'"
-                Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Validating the move"
+                #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Validating the move"
+                _verbose -message $strings.ValidateMove
                 Write-Debug $splat.query
                 Try {
                     $archived = Invoke-MySQLiteQuery @splat
                 }
                 Catch {
-                    Write-Warning "Failed to execute query $($splat.query)"
+                    Write-Warning ($strings.FailedQuery -f $splat.query)
+                    #"Failed to execute query $($splat.query)"
                     Close-MySQLiteDB $conn
                     Throw $_
                 }
                 if ($archived.taskid -eq $task.taskId) {
                     #remove the task from the tasks table
                     $splat.query = "DELETE from tasks WHERE taskid = '$($task.taskid)'"
-                    Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Removing the original task"
+                    #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Removing the original task"
+                    _verbose -message $strings.RemoveTask
                     Write-Debug $splat.query
                     Try {
                         Invoke-MySQLiteQuery @splat
                     }
                     Catch {
-                        Write-Warning "Failed to execute query $($splat.query)"
+                        Write-Warning ($strings.FailedQuery -f $splat.query)
+                        #"Failed to execute query $($splat.query)"
                         Close-MySQLiteDB $conn
                         Throw $_
                     }
                 }
                 else {
-                    Write-Warning "Could not verify that task $ID [$($task.taskid)] was copied to the archive table."
+                    Write-Warning ($strings.FailedVerifyTaskID -f $id,$task.taskID)
+                    # "Could not verify that task $ID [$($task.taskid)] was copied to the archive table."
                 }
                 if ($PassThru) {
                     $splat.Query = "SELECT * from archive WHERE taskid='$($task.taskid)'"
@@ -136,7 +153,8 @@ Function Complete-PSWorkItem {
                         _newWorkItemArchive $pass -path $Path
                     }
                     Catch {
-                        Write-Warning "Failed to execute query $($splat.query)"
+                        Write-Warning ($strings.FailedQuery -f $splat.query)
+                        #"Failed to execute query $($splat.query)"
                         Close-MySQLiteDB $conn
                         Throw $_
                     }
@@ -144,16 +162,19 @@ Function Complete-PSWorkItem {
             } #WhatIf
         } #if ID verified
         else {
-            Write-Warning "Failed to find task with id $ID"
+            Write-Warning ($strings.FailedToFind -f $id)
+            #"Failed to find task with id $ID"
         }
     } #process
 
     End {
+        $PSDefaultParameterValues["_verbose:block"] = "End"
         if ($conn.state -eq 'Open') {
-            Write-Verbose "[$((Get-Date).TimeOfDay) END    ] $($MyInvocation.MyCommand): Closing database connection."
+            #Write-Verbose "[$((Get-Date).TimeOfDay) END    ] $($MyInvocation.MyCommand): Closing database connection."
+            _verbose -message ($strings.CloseDBConnection -f $path)
             Close-MySQLiteDB -Connection $conn
         }
-        Write-Verbose "[$((Get-Date).TimeOfDay) END    ] $($MyInvocation.MyCommand): Ending "
+        _verbose -message $strings.Ending
+        #Write-Verbose "[$((Get-Date).TimeOfDay) END    ] $($MyInvocation.MyCommand): Ending "
     } #end
-
 }