+++ functions/public/Get-PSWorkItem.ps1
@@ -5,7 +5,7 @@ Function Get-PSWorkItem {
     Param(
         [Parameter(
             Position = 0,
-            HelpMessage = "The name of the work item. Wilcards are supported.",
+            HelpMessage = "The name of the work item. Wildcards are supported.",
             ValueFromPipelineByPropertyName,
             ParameterSetName = "name"
         )]
@@ -58,13 +58,19 @@ Function Get-PSWorkItem {
         [String]$Path = $PSWorkItemPath
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand): Starting "
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand): Detected culture $(Get-Culture)"
+        $PSDefaultParameterValues["_verbose:Command"] = $MyInvocation.MyCommand
+        $PSDefaultParameterValues["_verbose:block"] = "Begin"
+        _verbose -message $strings.Starting
+        #Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand): Starting "
+        #Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand): Detected culture $(Get-Culture)"
+        _verbose ($strings.DetectedCulture -f (Get-Culture))
     } #begin
 
     Process {
+        $PSDefaultParameterValues["_verbose:block"] = "Process"
         $results = [System.Collections.Generic.list[object]]::new()
-        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Detected parameter set $($PSCmdlet.parametersetname)"
+        #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Detected parameter set $($PSCmdlet.ParameterSetName)"
+        _verbose -message ($strings.DetectedParameterSet -f $PSCmdlet.ParameterSetName)
         Switch -regex ($PSCmdlet.ParameterSetName) {
             "all|days" { $query = "Select * from tasks" }
             "category" { $query = "Select * from tasks where category ='$Category' collate nocase" }
@@ -76,7 +82,7 @@ Function Get-PSWorkItem {
                 "days" {
                         $d = (Get-Date).AddDays($DaysDue)
                         $query = "Select * from tasks where duedate <= '$d' collate nocase"
-                    } 
+                    }
             #>
             "id" { $query = "Select * from tasks where ID ='$ID'" }
             "name" {
@@ -84,46 +90,53 @@ Function Get-PSWorkItem {
                     $Name = $name.replace("*", "%")
                 }
                 $query = "Select * from tasks where name = '$Name' collate nocase"
-
             }
         }
 
-        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): $query"
+        #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): $query"
+        _verbose -message $query
         $tasks = Invoke-MySQLiteQuery -Query $query -Path $Path
 
         if ($tasks.count -gt 0) {
-
-            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Found $($tasks.count) matching tasks"
+            #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Found $($tasks.count) matching tasks"
+            _verbose -message ($strings.FoundMatching -f $tasks.count)
             if ($PSCmdlet.ParameterSetName -eq 'days') {
                 $d = (Get-Date).AddDays($DaysDue)
-                Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Re-filtering for tasks due in the next $DaysDue day(s)."
-                Write-Verbose ("[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Cutoff date is {0}" -f $d)
-                Write-Verbose ("[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Filtering for items due before {0}" -f $d)
+                #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Re-filtering for tasks due in the next $DaysDue day(s)."
+                _verbose -message ($strings.RefilteringTasks -f $DaysDue)
+                #Write-Verbose ("[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Cutoff date is {0}" -f $d)
+                _verbose -message ($strings.CutOffDate -f $d)
+                #Write-Verbose ("[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Filtering for items due before {0}" -f $d)
+                _verbose -message ($strings.FilterItemsDue -f $d)
                 <#
                 This format doesn't appear to respect culture
                 $tasks = ($tasks).Where({ [DateTime]$_.duedate -le $d })
                 #>
                 $tasks = $tasks | Where-Object { ($_.duedate -as [DateTime]) -le $d}
-                Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Re-Filtering found $($tasks.count) items"
+                #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Re-Filtering found $($tasks.count) items"
+                _verbose -message ($strings.Refiltering -f $tasks.count)
             }
             $i=0
             foreach ($task in $tasks) {
-              Write-Debug "Converting id $($task.id)"
-              $nwi = _newWorkItem $task -path $path
-              write-Debug "Adding $($nwi.name) to the result list"
+            Write-Debug "Converting id $($task.id)"
+                $nwi = _newWorkItem $task -path $path
+                Write-Debug "Adding $($nwi.name) to the result list"
                 $results.Add($nwi)
                 $i++
-            }
+            } #foreach
             Write-Debug "processed $i tasks"
-            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Sorting $($results.count) results"
+            #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Sorting $($results.count) results"
+            _verbose -message ($strings.Sorting -f $results.count)
             $results | Sort-Object -Property DueDate
         }
         else {
-            Write-Warning "Failed to find any matching tasks"
+            Write-Warning $strings.WarnNoTasksFound
         }
     } #process
 
     End {
-        Write-Verbose "[$((Get-Date).TimeOfDay) END    ] $($MyInvocation.MyCommand): Ending"
+        $PSDefaultParameterValues["_verbose:block"] = "End"
+        _verbose -message $strings.Ending
+        #Write-Verbose "[$((Get-Date).TimeOfDay) END    ] $($MyInvocation.MyCommand): Ending"
     } #end
 }