+++ functions/public/Get-PSWorkItemDatabase.ps1
@@ -17,12 +17,19 @@ Function Get-PSWorkItemDatabase {
         [String]$Path = $PSWorkItemPath
     )
     Begin {
-        Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand): Starting "
+        $PSDefaultParameterValues["_verbose:Command"] = $MyInvocation.MyCommand
+        $PSDefaultParameterValues["_verbose:block"] = "Begin"
+        # Write-Verbose "[$((Get-Date).TimeOfDay) BEGIN  ] $($MyInvocation.MyCommand): $($strings.Starting)"
+        #_verbose -block begin -message $strings.testing
+        _verbose -message $strings.Starting
+        _verbose -message ($strings.UsingDB -f $path)
     } #begin
 
     Process {
-        Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Getting database information from $Path"
-
+        $PSDefaultParameterValues["_verbose:block"] = "Process"
+        #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Getting database information from $Path"
+        # Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): $($strings.getData) from $Path"
+        _verbose -message ($strings.GetData -f $Path)
         Try {
             $db = Get-MySQLiteDB -Path $Path -ErrorAction Stop
         }
@@ -30,18 +37,23 @@ Function Get-PSWorkItemDatabase {
             Throw $_
         }
         if ($db) {
-            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Opening a database connection"
+            #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): $($strings.OpenDBConnection)"
+            _verbose -message $strings.OpenDBConnection
             $conn = Open-MySQLiteDB -Path $path
-            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Getting task count"
+            #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): $($strings.TaskCount)"
+            _verbose -message $strings.TaskCount
             $tasks = Invoke-MySQLiteQuery -Connection $conn -KeepAlive -Query "Select Count() from tasks" -ErrorAction Stop
-            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Getting archive count"
+            #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): $($strings.ArchiveCount)"
+            _verbose -message $strings.ArchiveCount
             $archived = Invoke-MySQLiteQuery -Connection $conn -KeepAlive -Query "Select Count() from archive" -ErrorAction Stop
-            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Getting category count"
+            #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): $($strings.CategoryCount)"
+            _verbose -message $strings.CategoryCount
             $categories = Invoke-MySQLiteQuery -Connection $conn -KeepAlive -Query "Select Count() from categories" -ErrorAction Stop
-            Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): Closing the database connection"
+            #Write-Verbose "[$((Get-Date).TimeOfDay) PROCESS] $($MyInvocation.MyCommand): $($strings.CloseDBConnection)"
+            _verbose -message $strings.CloseDBConnection
             Close-MySQLiteDB -Connection $conn
 
-            #create a new PSWorkItemDatabase object
+            #create a new PSWorkItemDatabase object from the class definition
             $out = [PSWorkItemDatabase]::new()
             #define properties
             $out.Path = $db.Path
@@ -55,11 +67,14 @@ Function Get-PSWorkItemDatabase {
             $out.PageCount = $db.PageCount
             $out.PageSize = $db.PageSize
         }
+        #write the object to the pipeline
         $out
     } #process
 
     End {
-        Write-Verbose "[$((Get-Date).TimeOfDay) END    ] $($MyInvocation.MyCommand): Ending"
+        $PSDefaultParameterValues["_verbose:block"] = "End"
+        _verbose -message $strings.Ending
+        #Write-Verbose "[$((Get-Date).TimeOfDay) END    ] $($MyInvocation.MyCommand): $($strings.Ending)"
     } #end
 
 } #close Get-PSWorkItemDatabase